// src/shared/generated/BrowserRuntimeSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserRuntimeDisplayRecord } from "./BrowserRuntimeDisplayRecord";
import type { BrowserRuntimeRoleRecord } from "./BrowserRuntimeRoleRecord";
import type { BrowserRuntimeTabRecord } from "./BrowserRuntimeTabRecord";
import type { BrowserRuntimeWorkspaceRecord } from "./BrowserRuntimeWorkspaceRecord";

export type BrowserRuntimeSnapshot = { displays: Array<BrowserRuntimeDisplayRecord>, roles: Array<BrowserRuntimeRoleRecord>, tabs: Array<BrowserRuntimeTabRecord>, workspaces: Array<BrowserRuntimeWorkspaceRecord>, };