// src/shared/generated/RolePathsRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RolePathsRecord = { browserUserDataDir: string, systemBrowserDataDir: string, webview2UserDataDir: string, chromiumUserDataDir: string, webkitDataStoreKey: string, webkitDataStoreIdentifier: string, };